Changed shortcut to alternative for consistency
I changed the shortcut for open file in (horizontal) split to `<c-s>`. I think it is much easier to memorize: `s` for the default (horizontal) `:split` command.

Also I changed the order of the shortcuts because `<c-s>` is the "default" horizontal split vs. `:vs[plit]`, `:vert[ical] sp[lit]`

The consistency to the commands in the Vim quickfix list is not fully perfect, though:

 - `<c-e>` -> `e` -> edit
 - `<c-t>` -> `t` -> new tab
 - `<c-v>` -> `v` -> vertical split
 - `<c-s>` -> `h` -> (horizontal) split, like `:split`